refactor(resolver): split into modules
