+++ crates/keyton_rust_compiler/src/shir/resolver/mod.rs
@@ -1,0 +1,11 @@
+mod errors;
+mod core;
+mod defs;
+mod stmt;
+mod expr;
+mod user_funcs;
+mod program;
+
+pub use core::Resolver;
+pub use errors::{ResolveError, ResolveReport};
+pub use program::{ResolvedProgram, resolve_program, resolve_program_with_spans};
